+++ .gitignore
@@ -13,6 +13,7 @@ __pycache__/
 .venv
 .idea
 .env
+MagicMock/
 .ipynb_checkpoints
 .ropeproject
 Untitled.ipynb
