--- tests/00initial/test_logging.py
@@ -1,295 +1,0 @@
-import logging  # Added
-import uuid
-from io import StringIO  # Added
-from unittest.mock import MagicMock, patch  # Added
-
-import pytest
-
-from buttermilk._core.bm_init import BM  # Modified import
-from buttermilk._core.config import CloudProviderCfg  # Added for mocking
-from buttermilk._core.context import set_logging_context  # Added
-from buttermilk._core.log import logger
-
-DEBUG_TEXT = "this should not show up in the log" + str(uuid.uuid1())
-LOG_TEXT = "logging appears to be working" + str(uuid.uuid1())
-
-# bm instance from dmrc is used by default by some tests.
-# bm_instance = bm
-
-
-@pytest.fixture(scope="function")
-def bm_instance(tmp_path) -> BM:
-    """Provides a BM instance with a temporary save directory."""
-    from buttermilk._core.bm_init import create_session_bm
-    
-    # Create a session-scoped BM instance for testing
-    test_bm = create_session_bm(
-        name="test_bm_instance",
-        job="test_job",
-        platform="test",
-        save_dir_base=str(tmp_path),  # Use pytest's tmp_path for a unique temp dir
-        # No cloud infrastructure by default for most tests
-        cloud_manager=None,
-        secret_manager=None,
-        llms_instance=None,
-        logger_cfg=None,  # Disable cloud logging by default for most tests
-    )
-    return test_bm
-
-
-@pytest.fixture(scope="function")
-def configured_logger(bm_instance):
-    # In the new session-scoped architecture, logging is set up automatically
-    # during BM initialization. We just need to set up console/file logging for tests.
-    from buttermilk._core.log import setup_console_logging
-    
-    # Set up console logging for test visibility
-    setup_console_logging(verbose=True)
-    
-    # The global buttermilk logger is ready to use
-    yield logger
-    
-    # Cleanup: remove handlers to avoid test interference
-    import logging as std_logging
-    root_logger = std_logging.getLogger()
-    for handler in root_logger.handlers[:]:
-        root_logger.removeHandler(handler)
-    
-    # Reset context vars
-    set_logging_context(None, None, None)
-
-
-# Test Case 1: Context Variables Set
-def test_logging_with_context_vars(configured_logger):
-    test_session_id = "test_session_123"
-    test_agent_id = "test_agent_abc"
-    set_logging_context(test_session_id, test_agent_id)
-
-    log_buffer = StringIO()
-    # The format string should match the one in bm_init.py for console output
-    # console_format = "%(asctime)s %(hostname)s %(name)s [%(session_id)s:%(agent_id)s] %(filename)s:%(lineno)d %(levelname)s %(message)s"
-    # For simplicity in testing, we can use a more minimal format string that includes the context vars
-    test_formatter = logging.Formatter("[%(session_id)s:%(agent_id)s] %(message)s")
-
-    stream_handler = logging.StreamHandler(log_buffer)
-    stream_handler.setFormatter(test_formatter)
-
-    # Add handler to the global 'buttermilk' logger
-    configured_logger.addHandler(stream_handler)
-    # Ensure test messages go through by setting level if necessary (already set by configured_logger fixture)
-    # configured_logger.setLevel(logging.INFO)
-
-    log_message = "Test message with context"
-    configured_logger.info(log_message)
-
-    # Remove the handler to prevent interference and ensure log_buffer is flushed
-    configured_logger.removeHandler(stream_handler)
-    stream_handler.close()
-
-    log_output = log_buffer.getvalue()
-
-    assert f"[{test_session_id}:{test_agent_id}] {log_message}" in log_output
-    # Cleanup context variables
-    set_logging_context(None, None)
-
-
-# Test Case 2: Context Variables Not Set (Default)
-def test_logging_without_context_vars(configured_logger):
-    set_logging_context(None, None)  # Ensure defaults
-
-    log_buffer = StringIO()
-    test_formatter = logging.Formatter("[%(session_id)s:%(agent_id)s] %(message)s")
-    stream_handler = logging.StreamHandler(log_buffer)
-    stream_handler.setFormatter(test_formatter)
-    configured_logger.addHandler(stream_handler)
-    # configured_logger.setLevel(logging.INFO)
-
-    log_message = "Test message without context"
-    configured_logger.info(log_message)
-
-    configured_logger.removeHandler(stream_handler)
-    stream_handler.close()
-
-    log_output = log_buffer.getvalue()
-    # Default None values for context vars should appear as "None" in the string
-    assert f"[None:None] {log_message}" in log_output
-
-
-# Test Case 3: Cloud Logging with Session-Scoped Architecture
-@patch("google.cloud.logging_v2.handlers.CloudLoggingHandler")  # Patch the actual class
-def test_cloud_logging_with_session_context(MockCloudLoggingHandler, tmp_path):
-    from buttermilk._core.bm_init import create_session_bm
-    from buttermilk._core.config import LoggerConfig
-    from buttermilk._core.cloud import CloudManager
-    
-    # Create a mock logger configuration
-    mock_logger_cfg = LoggerConfig(
-        type="gcp", 
-        project_id="test-gcp-project", 
-        location="us-central1"
-    )
-
-    # Mock cloud manager and GCS log client
-    mock_cloud_manager = MagicMock()
-    mock_log_client = MagicMock()
-    mock_cloud_manager.gcs_log_client.return_value = mock_log_client
-
-    # Create session-scoped BM instance with cloud logging enabled
-    cloud_bm = create_session_bm(
-        name="test_cloud_bm",
-        job="test_cloud_job",
-        platform="test",
-        save_dir_base=str(tmp_path),
-        cloud_manager=mock_cloud_manager,
-        logger_cfg=mock_logger_cfg,  # Enable cloud logging
-    )
-
-    # The mock CloudLoggingHandler instance will be created during BM initialization
-    mock_handler_instance = MockCloudLoggingHandler.return_value
-    mock_handler_instance.emit = MagicMock()  # Mock the emit method
-    mock_handler_instance.handle = MagicMock()  # Also mock handle as it's often called
-
-    # In the new architecture, cloud logging is set up automatically during BM initialization
-    # The session context is already established with the real session_id from the BM
-    
-    test_session_id = cloud_bm.session_info.session_id  # Use the real session ID
-    test_agent_id = "cloud_agent_xyz"
-    set_logging_context(test_session_id, None, test_agent_id)  # Update context with agent
-
-    log_message = "Test cloud message"
-    # Use the global logger instance
-    global_logger = logging.getLogger("buttermilk")
-    global_logger.info(log_message)
-
-    # Assertions
-    MockCloudLoggingHandler.assert_called_once()  # Check handler was instantiated
-
-    # Check that emit or handle was called on the handler instance
-    # The actual method called might depend on the logging library's internals
-    # For standard library logging, `handle` is usually called, which then calls `emit`
-    # We check if *any* record passed to `handle` has the correct attributes
-    call_args_list = mock_handler_instance.handle.call_args_list
-    assert len(call_args_list) > 0, "CloudLoggingHandler.handle was not called"
-
-    record_found = False
-    for call_args in call_args_list:
-        record = call_args[0][0]  # The first argument to handle is the LogRecord
-        if hasattr(record, "session_id") and record.session_id == test_session_id and \
-           hasattr(record, "agent_id") and record.agent_id == test_agent_id and \
-           record.getMessage() == log_message:
-            record_found = True
-            break
-
-    assert record_found, "LogRecord with correct context and message not found in CloudLoggingHandler calls"
-
-    # Cleanup
-    set_logging_context(None, None)
-    # Clean up handlers from the global logger
-    for handler in global_logger.handlers[:]:
-        global_logger.removeHandler(handler)
-
-
-# Keep existing tests if they are still relevant and don't conflict.
-# The bm fixture might need to be the new bm_instance fixture.
-# capsys might still be useful for some tests, but new tests use StringIO.
-
-@pytest.fixture(scope="function")
-def logger_new(bm_instance):  # Use the new bm_instance fixture
-    from buttermilk._core.log import setup_console_logging
-    
-    # Set up console logging for test visibility (session logging is automatic)
-    setup_console_logging(verbose=True)
-    
-    # logger instance is the global buttermilk logger
-    yield logger
-    logger.info("Tearing test logger_new down.")
-    
-    # It's important to clean up handlers to prevent test interference
-    import logging as std_logging
-    root_logger = std_logging.getLogger()
-    for handler in root_logger.handlers[:]:
-        root_logger.removeHandler(handler)
-
-
-def test_error(capsys, logger_new):  # logger_new now uses bm_instance
-    log_text_error = f"{LOG_TEXT}_error_{uuid.uuid4()}"
-    logger_new.error(log_text_error)
-    captured = capsys.readouterr()
-    assert log_text_error in captured.err
-    assert "error" in str.lower(captured.err)
-
-
-@pytest.mark.anyio
-async def test_warning(capsys, logger_new, bm_instance: BM):  # logger_new and bm_instance
-    log_text_warning = f"{LOG_TEXT}_warning_{uuid.uuid4()}"
-    logger_new.warning(log_text_warning)
-    captured = capsys.readouterr()
-    assert log_text_warning in captured.err
-    assert "warning" in str.lower(captured.err)
-
-    # The cloud logging part of this test might be complex with the new bm_instance
-    # if bm_instance is not configured for cloud logging by default.
-    # For now, let's assume bm_instance is NOT doing cloud logging unless specified.
-    # If cloud logging needs to be tested here, bm_instance needs logger_cfg.
-    # This part of the test is commented out as it requires a bm_instance
-    # specifically configured for cloud logging, and the test_cloud_logging_with_context_vars
-    # now covers specific cloud logging mocking.
-
-    # await asyncio.sleep(5)
-    # from google.cloud.logging_v2 import DESCENDING
-    # entries = bm_instance.gcs_log_client.list_entries( # This would fail if gcs_log_client is not set up
-    #     order_by=DESCENDING,
-    #     max_results=100,
-    # )
-    # for entry in entries:
-    #     if log_text_warning in str(entry.payload):
-    #         return True
-    # raise OSError(f"Warning message not found in log: {log_text_warning}")
-
-
-def test_debug(capsys, logger_new):  # logger_new uses bm_instance
-    # bm_instance in configured_logger fixture sets verbose=True for setup_logging
-    # So, debug messages should be processed by the logger.
-    # Whether they appear in capsys.err depends on the console handler's level.
-    # The default coloredlogs setup in BM sets console to DEBUG if verbose is True.
-    debug_text_specific = f"{DEBUG_TEXT}_debug_{uuid.uuid4()}"
-    logger_new.debug(debug_text_specific)
-    captured = capsys.readouterr()
-    # If verbose=True was used in setup_logging for logger_new, debug messages should appear.
-    assert debug_text_specific in captured.err
-    assert "debug" in str.lower(captured.err)
-
-
-# This fixture might need to change if bm_instance doesn't always have gcs_log_client
-# @pytest.fixture
-# def cloud_logging_client_gcs(bm_instance: BM):
-#     # This assumes bm_instance is configured with GCP logging.
-#     # If not, this fixture will fail.
-#     if not (bm_instance.logger_cfg and bm_instance.logger_cfg.type == "gcp"):
-#         pytest.skip("Skipping GCS log client test as BM instance is not configured for GCP logging")
-#     return bm_instance.gcs_log_client # This might error if not configured
-
-
-def test_info(capsys, logger_new):  # logger_new uses bm_instance
-    log_text_info = f"{LOG_TEXT}_info_{uuid.uuid4()}"
-    logger_new.info(log_text_info)
-    captured = capsys.readouterr()
-    assert log_text_info in captured.err
-    assert "info" in str.lower(captured.err)
-
-
-# This test is problematic because:
-# 1. It relies on 'google.cloud.logging' which wasn't explicitly imported.
-# 2. It assumes cloud_logging_client_gcs is available and working.
-# 3. It checks for DEBUG_TEXT which might have been logged by other tests if logging state persists.
-# Given the new mocking test for cloud, this might be redundant or need heavy refactoring.
-# For now, commenting out to focus on the new context-based tests.
-
-# def test_cloud_loger_debug(cloud_logging_client_gcs):
-#     entries = cloud_logging_client_gcs.list_entries(
-#         order_by=google.cloud.logging.DESCENDING, # Needs import: from google.cloud import logging as google_logging
-#         max_results=100,
-#     )
-#     for entry in entries:
-#         if DEBUG_TEXT in str(entry.payload): # DEBUG_TEXT is a global, could be from anywhere
-#             raise OSError(f"Debug message found in log: {LOG_TEXT}")
