+++ tests/integration/00initial/test_bm_singleton.py
