+++ tests/integration/00initial/test_setup.py
@@ -1,0 +1,39 @@
+import os
+import sys
+
+import google.auth
+import google.auth.credentials
+import pytest
+
+from buttermilk import BM, logger
+from buttermilk._core.log import logger  # noqa
+
+
+class Test00Setup:
+    def test_imports(self):
+        pass
+
+    def test_python_version(self):
+        """Check that the Python version is 3.10 or higher."""
+        assert sys.version_info >= (3, 10)
+
+    def test_gcloud_credentials_adc(self):
+        credentials, project_id = google.auth.default()
+        assert credentials
+        assert project_id
+
+    def test_gcloud_no_json_key(self):
+        """Check that the JSON key is not set."""
+        assert "GOOGLE_APPLICATION_CREDENTIALS" not in os.environ
+
+    def test_bigquery(self):
+        """Test moved to tests/integration/test_setup_infrastructure.py"""
+        pytest.skip("Moved to integration tests - see test_setup_infrastructure.py")
+
+    def test_database(self):
+        """Test moved to tests/integration/test_setup_infrastructure.py"""
+        pytest.skip("Moved to integration tests - see test_setup_infrastructure.py")
+
+    def test_hf_login(self):
+        """Test moved to tests/integration/test_setup_infrastructure.py"""
+        pytest.skip("Moved to integration tests - see test_setup_infrastructure.py")
