+++ tests/integration/40apis/test_gemini_vertex_embeddings_integration.py
