+++ tests/integration/test_azure.py
